Refactor core files and add first tests
