+++ .gitignore
@@ -1,3 +1,6 @@
 target
 .DS_Store
 .vscode
+.env
+coveralls*
+*.prof
